refactor: page selection
